+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - CHANGED: Renamed Whois::Record::Scanners::Ast to Renamed Whois::Record::Scanners::Nodable
 
+- CHANGED: Updated whois.nic.af parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).